Added graphics window, renderer, and entity
This is the base commit for the program. From here, we can add entities, adding them
to the rendering queue, and they should get drawn on each game loop. I am hoping that
there aren't concurrent modification exceptions. Ideally, we'll run the game on multiple
threads so that rendering and updates are kept separate. I have no idea how 'expensive'
this game will be.