Key presses sucked with screen transitions. Now fixed.
I have changed things up a bit. The pause function is still controlled by
the ESCAPE key within the game, but to continue the game or exit to the main
menu, you have to use the buttons in the pause screen. I was encountering
problems with key conflict. If the ESCAPE key in the game paused the game,
and the ESCAPE key also exited to the main menu, pausing the game would take
you straight to the main menu without going through the pause screen (because
the transition was so fast, the key didn't disable). Now it works just fine.