Added Safezone collision stuffs and multithreading (no synchronization yet)
I thought that the Safezones were taking their time to update because of
lag...so I implemented multithreading. This is very simply placed now, I just
named the threads "draw" and "update", and check which thread is causing the run()
function to be called to determine whether or not I need to draw or update the
game. This, however, didn't solve the issue. The issue arose when I wasn't checking
if the Safezone it was intersecting already had another Safezone stored in its
Intersections arraylist, so I was adding duplicates such that when they were no longer
intersecting, the Safezone still thought that they were, so it didn't update the color.
It is working now. The color mixing isn't complete yet (I don't know if it works with
a mass number of Safezones yet) but it will be refined.