Safezone colors mix reliably and cores spawn in random locations on the screen
With this change, colors mix reliably now, but I have introduced potential for
concurrent modification exceptions. The number of comparisons that have to happen
to merge a long chain of safezones take longer than the game loop, so safezones are
colliding with other safezones and adding them to their intersections list while
the color mix is occuring. We can't synchronize arraylist objects, but we can definitely
put blockers on them. I'm not familiar with the concept of Semaphores yet to know
conclusively if it will solve our problem, but it might.