Fixed concurrency issues
In Safezone, SafezoneManager, and CoreManager concurrent modifications were at risk
of occuring. I resolved these by replacing all "thread unsafe" ArrayLists with
CopyOnWriteArrayLists. CopyOnWriteArrayLists complete all actions on a thread
prior to the accessing or removal of elements on another thread (using a
happens-before queue of sorts). This seems to be the best concurrency solution we
can get right off the shelf without having to put in any significant effort.