Music now stops when you leave the main menu
I made the clip in the main menu a global private variable and modified
the setVisible function to pause and play the music depending on the
visibility of the JPanel. This is a temporary solution, and I am wondering
if there is a way to fade out music. We might be able to control volume
on the threads that run the music. Just a thought.